useeffect added to pass first suggestion to search widget when search results are empty
added suggest-complete event handler to pass suggestion results to searchSuggestions variable. When the search-complete event results are empty the first search term from the search suggestion event results will be bassed to the search widget.

TODO: handling updates to search term in the search bar and the url parameter. Currently the search bar text and url paramter update to the fist suggestion instead of the original search term from the url parameter.